requirements.txt
# Python
python-dotenv>=1.0.1
requests>=2.32.3

# Jupyter Notebook and Kernel
ipykernel>=6.29.5
ipywidgets>=8.1.5
notebook>=7.2.2

# FastAPI
aiomysql>=0.2.0
alembic>=1.13.2
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35

# Hugging Face Dependencies
# accelerate>=0.34.2
# huggingface_hub>=0.24.6
# numpy<2
# torch>=2.2.2
# transformers>=4.44.2
